requirements.txt

flask
pyjwt
